refactor(adapter): one node:http external rewrite proxy shared by edge and function
Deno ships node:http/https too and its client passes compressed bodies through
raw (verified), so the edge handler no longer needs the fetch() variant that had
to drop content-encoding. Lives in adapter-runtime-shared next to the routing
helpers; both handlers just call proxyExternalRewrite().
